move twisted edward curve test cases to instances
